\name{RWOAG-package}
\alias{RWOAG-package}
\alias{RWOAG}
\docType{package}
\title{
Performs random walk on a graph
}
\description{
Given an adjacency matrix or graphNEL object, a random walk is performed.
Returns vector of all nodes in the matrix, and their visitation probability from the random walk
}
\details{
\tabular{ll}{
Package: \tab RWOAG\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-06-07\cr
License: \tab LPGL-3\cr
LazyLoad: \tab yes\cr
}
library(RWOAG)
}
\author{
Written by James Perkins and Ian Morilla
Maintainer: James Perkins <jperkins@biochem.ucl.ac.uk>
}
\references{
Kohler et al., 2007 (pmid:18371930)
}
\examples{
## Examples given in the man page for randomWalk
?randomWalk
}
